updated the models
